+++ vue/src/store/index.js
@@ -1,15 +1,11 @@
-import Vue from 'vue';
-import Vuex from 'vuex';
-import state from './state';
-import mutations from './mutations';
-import actions from './actions';
-import getters from './getters';
+import Vue from 'vue'
+import Vuex from 'vuex'
+import game from './game';
 
 Vue.use(Vuex);
 
 export default new Vuex.Store({
-    state,
-    mutations,
-    getters,
-    actions,
+    modules: {
+        game
+    }
 })
